+++ modules/bridge/src/main/ts/ephox/bridge/api/PublicDialog.ts
@@ -15,7 +15,7 @@ import { DropZoneSpec } from '../components/dialog/Dropzone';
 import { GridSpec } from '../components/dialog/Grid';
 import { HtmlPanelSpec } from '../components/dialog/HtmlPanel';
 import { IframeSpec } from '../components/dialog/Iframe';
-import { ImagePanelSpec } from '../components/dialog/ImagePanel';
+import { ImagePreviewSpec } from '../components/dialog/ImagePreview';
 import { InputSpec } from '../components/dialog/Input';
 import { LabelSpec } from '../components/dialog/Label';
 import { ListBoxItemSpec, ListBoxSpec, ListBoxSingleItemSpec, ListBoxNestedItemSpec } from '../components/dialog/ListBox';
@@ -71,7 +71,7 @@ export {
 
   IframeSpec,
 
-  ImagePanelSpec,
+  ImagePreviewSpec as ImagePanelSpec,
 
   InputSpec,