+++ modules/bridge/src/main/ts/ephox/bridge/components/dialog/ImagePreview.ts
@@ -3,14 +3,14 @@ import { Optional } from '@ephox/katamari';
 
 import { FormComponent, formComponentFields, FormComponentSpec } from './FormComponent';
 
-export interface ImagePanelSpec extends FormComponentSpec {
-  type: 'imagepanel';
+export interface ImagePreviewSpec extends FormComponentSpec {
+  type: 'imagepreview';
   width?: string;
   height?: string;
 }
 
-export interface ImagePanel extends FormComponent {
-  type: 'imagepanel';
+export interface ImagePreview extends FormComponent {
+  type: 'imagepreview';
   width: Optional<string>;
   height: Optional<string>;
 }